add basic check_session layout andconnect to account store
In order to sensibly handle session checking in the router, the page loading animation part of the componentWillMount lifecycle function had to be broken out into its own function (fadein_animation())

Also, added the basic template for the check_session() function that will look for a user cookie in the browser, then use that to log into the server and handle the redux account state.

In order to POST login I preemptively added the dispatch event to the component and added a mapDispatchToProps object.

Lastly, docStrings.