logout user when the login page mounts
Any time a user tries to access something which is locked by connect-ensure-login on the server, they are brought to the login page. The login page completely clears the authentication status client side, since by accessing the login route it is proven that the user is not authenticated on the server.